refactor: remove  package and implement BuildCLICommand interface
