+++ cli/src/main/java/dev/buildcli/cli/BuildCLI.java
@@ -2,8 +2,7 @@ package dev.buildcli.cli;
 
 import dev.buildcli.cli.commands.*;
 import dev.buildcli.cli.commands.AiCommand;
-import dev.buildcli.cli.commands.bug.BugCommand;
-import picocli.CommandLine;
+import dev.buildcli.cli.commands.BugCommand;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.HelpCommand;
